Fixed final page Added "Finish" button

+++ src/NotificationCenter.cpp
@@ -15,7 +15,7 @@ namespace SubutaiLauncher
 
     void NotificationCenter::add(ScriptAction action)
     {
-        _logger->debug("Notification received: %d", action);
+        _logger->debug("Notification received: %d", (int)action);
         _pool.push_back(action);
     }
